Generate code for local calls
This commit introduces substantial changes how we will work with method
calls
Instead of relying on unconditional jumps, we will introduce call and
ret statements.
This streamlines code generation to some extend as a function does not
need to know from where exactly it was called. It puts additional work
on the VM, however, to keep track of the call stack and return
addresses.